+++ aioxmpp/__init__.py
@@ -101,6 +101,7 @@ from .shim import SHIMService  # NOQA
 from .adhoc import AdHocClient, AdHocServer  # NOQA
 from .avatar import AvatarClient, AvatarServer  # NOQA
 from .blocking import BlockingClient  # NOQA
+from .carbons import CarbonsClient  # NOQA
 
 
 def set_strict_mode():